fix based on requested changes
